requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
dacite==1.8.1
fonttools==4.42.0
gitdb==4.0.10
GitPython==3.1.32
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
multimethod==1.9.1
networkx==3.1
numba==0.57.1
numpy==1.23.5
packaging==23.1
pandas==2.0.3
pandas-profiling==3.6.6
patsy==0.5.3
phik==0.12.3
Pillow==9.5.0
plotly==5.16.1
protobuf==4.24.0
psycopg2-binary==2.9.7
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
shap==0.42.1
six==1.16.0
slicer==0.0.7
smmap==5.0.0
statsmodels==0.14.0
streamlit==1.25.0
streamlit-elements==0.1.0
streamlit-pandas-profiling==0.1.3
streamlit-shap==1.0.2
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
visions==0.7.5
wordcloud==1.9.2
xgboost==1.7.6
ydata-profiling==4.5.1
zipp==3.16.2
